fix test

+++ dot/core/digest_test.go
@@ -342,7 +342,7 @@ func TestDigestHandler_HandleBABEOnDisabled(t *testing.T) {
 
 	err = handler.HandleConsensusDigest(d, nil)
 	require.NoError(t, err)
-	require.Equal(t, uint64(7), handler.babe.(*mockBlockProducer).disabled)
+	require.Equal(t, uint32(7), handler.babe.(*mockBlockProducer).disabled)
 }
 
 func createHeaderWithPreDigest(slotNumber uint64) *types.Header {
